backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic
passlib[bcrypt]
python-jose[cryptography]
python-multipart
